i have changed the int to float

+++ if.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-	int num;
+	float num;
 	printf("enter the num\n");
 	scanf("%d",&num);
 	if(num%2==0)
